raft2: - remove old classes - rewrite streamable

+++ src/main/java/org/mitallast/queue/Version.java
@@ -3,19 +3,19 @@ package org.mitallast.queue;
 import java.io.IOException;
 
 public class Version {
-    public static final int V1_0_0_ID = 10000;
+    public static final short V1_0_0_ID = 10000;
     public static final Version V1_0_0 = new Version(V1_0_0_ID);
 
     public static final Version CURRENT = V1_0_0;
-    public static final int CURRENT_ID = V1_0_0_ID;
+    public static final short CURRENT_ID = V1_0_0_ID;
 
-    public final int id;
+    public final short id;
 
-    Version(int id) {
+    Version(short id) {
         this.id = id;
     }
 
-    public static Version fromId(int id) throws IOException {
+    public static Version fromId(short id) throws IOException {
         switch (id) {
             case CURRENT_ID:
                 return V1_0_0;
